requirements.txt
flask
python-telegram-bot
shazamio
yt-dlp
ffmpeg-python
